requirements.txt
aiohttp==3.7.4.post0
akinator.py==5.0.0
async-cleverbot==0.2.2
async-timeout==3.0.1
asyncdagpi==3.3.4
attrs==21.2.0
backoff==1.11.1
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
discord-ext-menus==1.1
discord-ext-owoify==0.0.7.9a0
discord.py @ git+https://github.com/Rapptz/discord.py.git@45d498c1b76deaf3b394d17ccf56112fa691d160
DiscordUtils==1.3.4
dislash.py==1.4.9
eight-ball==3.0.3
english-words==1.0.4
ffmpeg==1.4
greenlet==1.1.2
gTTS==2.2.3
idna==3.2
mal-api==0.4.0
multidict==5.1.0
mystbin.py==2.1.3
orjson==3.6.3
parsedatetime==2.6
Pillow==8.3.2
psutil==5.8.0
pycparser==2.20
pygit2==1.7.0
PyNaCl==1.4.0
pypresence==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
requests==2.26.0
sentry-sdk==1.3.1
six==1.16.0
soupsieve==2.2.1
spotify==0.10.2
SQLAlchemy==1.4.25
statcord.py==3.1.0
TenGiphPy==2.5.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.7.0
youtube-dl==2021.6.6
youtube-search==2.1.0
